open_uri modification
if there is several args (the first without ://) :
We USE (not just a test anymore) search engine.
Url doesn't have space, that's why if there is one we believe that
user want to use search engine.
But if there is absolutely no search engine we fill space with '%20'
and add a leading 'http://'.

Signed-off-by: Maxime de Roucy <maxime.deroucy@gmail.com>